Fix: Refactor date formatting to support localized month names and structured data models
This change replaces simple string-based date formatting with a more robust system using `DateModel` and `DatePresentationModel`. It adds localized month strings across multiple languages and introduces a `MonthPresentationMapper` to handle translation.

Key changes:
- Created `DateModel` (core) and `DatePresentationModel` (UI) for structured date handling.
- Refactored `ReadDateFormatter` to return `DatePresentationModel` instead of a raw string.
- Added localized month names to `strings.xml` for English, Spanish, and Brazilian Portuguese.
- Extracted the "edit" string and created a reusable `EditTextButton` in the UI component module.
- Modularized `CompletedDateSection` into smaller components like `CompletedDateCard` and `DateText`.
- Bumped version to 1.2.1 (versionCode 5).
