+++ feature/day/src/commonMain/kotlin/com/quare/bibleplanner/feature/day/presentation/component/DayReadSection.kt
@@ -12,9 +12,9 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.quare.bibleplanner.feature.day.presentation.component.completed_date.CompletedDateSection
-import com.quare.bibleplanner.ui.component.date.DatePresentationModel
+import com.quare.bibleplanner.feature.day.presentation.component.completeddate.CompletedDateSection
 import com.quare.bibleplanner.feature.day.presentation.model.DayUiEvent
+import com.quare.bibleplanner.ui.component.date.DatePresentationModel
 import kotlin.time.ExperimentalTime
 
 @OptIn(ExperimentalTime::class, ExperimentalMaterial3Api::class)
