+++ ui/component/src/commonMain/kotlin/com/quare/bibleplanner/ui/component/icon/EditTextButton.kt
@@ -11,11 +11,11 @@ import org.jetbrains.compose.resources.stringResource
 @Composable
 fun EditTextButton(
     modifier: Modifier = Modifier,
-    onClick: () -> Unit
+    onClick: () -> Unit,
 ) {
     TextButton(
         modifier = modifier,
-        onClick = onClick
+        onClick = onClick,
     ) {
         Text(stringResource(Res.string.edit))
     }
